# docs/swagger.yaml
basePath: /swagger/ebcp-service
definitions:
...
  common.PageGeneric-model_Ebcp_player_info:
    properties:
      items:
        items:
          $ref: '#/definitions/model.Ebcp_player_info'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  common.Response:
    properties:
      data: {}
      msg:
        type: string
      status:
        type: integer
    type: object
...
  model.Ebcp_player_info:
    properties:
      current_program_id:
        description: 当前节目ID
        type: string
      current_program_state:
        description: 当前节目状态,0:播放,1:暂停,2:停止
        type: integer
      exhibition_id:
        description: 所属展览ID
        type: string
      exhibition_name:
        description: 所属展览名称
        type: string
      id:
        description: 设备ID
        type: string
      ip_address:
        description: IP地址
        type: string
      item_id:
        description: 所属展项ID
        type: string
      item_name:
        description: 所属展项名称
        type: string
      item_type:
        description: 所属展项类型
        type: string
      name:
        description: 设备名称
        type: string
      port:
        description: 端口
        type: integer
      programs:
        description: 关联的节目信息
      room_floor:
        description: 所属展厅楼层
        type: string
      room_floor_name:
        description: 所属展厅楼层名称
        type: string
      room_floor_value:
        description: 所属展厅楼层值
        type: string
      room_id:
        description: 所属展厅ID
        type: string
      room_name:
        description: 所属展厅名称
        type: string
      sound_state:
        description: 音量状态,0:静音,1:非静音
        type: integer
      status:
        description: 状态,1:正常,2:离线,3:故障
        type: integer
      version:
        description: 版本
        type: string
      volume:
        description: 音量(0-100)
        type: integer
    type: object
...
info:
  contact: {}
  description: ebcp-service API
  title: ebcp-service API
  version: "1.0"
paths:
...
  /ebcp-player-info:
    get:
      description: query objects
      parameters:
      - description: _select
        in: query
        name: _select
        type: string
      - description: order
        in: query
        name: _order
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: ip_address
        in: query
        name: ip_address
        type: string
      - description: port
        in: query
        name: port
        type: string
      - description: version
        in: query
        name: version
        type: string
      - description: status
        in: query
        name: status
        type: string
      - description: current_program_id
        in: query
        name: current_program_id
        type: string
      - description: current_program_state
        in: query
        name: current_program_state
        type: string
      - description: volume
        in: query
        name: volume
        type: string
      - description: sound_state
        in: query
        name: sound_state
        type: string
      - description: item_id
        in: query
        name: item_id
        type: string
      - description: item_name
        in: query
        name: item_name
        type: string
      - description: item_type
        in: query
        name: item_type
        type: string
      - description: room_id
        in: query
        name: room_id
        type: string
      - description: room_name
        in: query
        name: room_name
        type: string
      - description: room_floor
        in: query
        name: room_floor
        type: string
      - description: room_floor_value
        in: query
        name: room_floor_value
        type: string
      - description: room_floor_name
        in: query
        name: room_floor_name
        type: string
      - description: exhibition_id
        in: query
        name: exhibition_id
        type: string
      - description: exhibition_name
        in: query
        name: exhibition_name
        type: string
      - description: programs
        in: query
        name: programs
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Ebcp_player_info'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: query objects
      tags:
      - 播放设备详细视图，包含设备信息及其关联的展项、展厅、展览和节目信息（JSON格式）
  /ebcp-player-info/page:
    get:
      description: page query, _page(from 1 begin), _page_size, _order, and others
        fields, status=1, name=$like.%CAM%
      parameters:
      - description: current page
        in: query
        name: _page
        required: true
        type: integer
      - description: page size
        in: query
        name: _page_size
        required: true
        type: integer
      - description: order
        in: query
        name: _order
        type: string
      - description: _select
        in: query
        name: _select
        required: true
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: ip_address
        in: query
        name: ip_address
        type: string
      - description: port
        in: query
        name: port
        type: string
      - description: version
        in: query
        name: version
        type: string
      - description: status
        in: query
        name: status
        type: string
      - description: current_program_id
        in: query
        name: current_program_id
        type: string
      - description: current_program_state
        in: query
        name: current_program_state
        type: string
      - description: volume
        in: query
        name: volume
        type: string
      - description: sound_state
        in: query
        name: sound_state
        type: string
      - description: item_id
        in: query
        name: item_id
        type: string
      - description: item_name
        in: query
        name: item_name
        type: string
      - description: item_type
        in: query
        name: item_type
        type: string
      - description: room_id
        in: query
        name: room_id
        type: string
      - description: room_name
        in: query
        name: room_name
        type: string
      - description: room_floor
        in: query
        name: room_floor
        type: string
      - description: room_floor_value
        in: query
        name: room_floor_value
        type: string
      - description: room_floor_name
        in: query
        name: room_floor_name
        type: string
      - description: exhibition_id
        in: query
        name: exhibition_id
        type: string
      - description: exhibition_name
        in: query
        name: exhibition_name
        type: string
      - description: programs
        in: query
        name: programs
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/common.PageGeneric-model_Ebcp_player_info'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: page query
      tags:
      - 播放设备详细视图，包含设备信息及其关联的展项、展厅、展览和节目信息（JSON格式）
...
swagger: "2.0"